Added some documentation.
